Replace metrics page with updated version
