--- src/DevOpsAssistant/DevOpsAssistant.Tests/Pages/MetricsBetaPageTests.cs
@@ -1,123 +1,0 @@
-using Bunit;
-using System;
-using System.Reflection;
-using System.Collections.Generic;
-using DevOpsAssistant.Pages;
-using DevOpsAssistant.Services.Models;
-using DevOpsAssistant.Tests.Utils;
-using MudBlazor;
-
-namespace DevOpsAssistant.Tests.Pages;
-
-public class MetricsBetaPageTests : ComponentTestBase
-{
-    [Fact]
-    public void MetricsBeta_Renders()
-    {
-        SetupServices(includeApi: true);
-
-        var cut = RenderWithProvider<TestMetricsBeta>();
-
-        Assert.Contains("Load", cut.Markup);
-        Assert.Contains("Export CSV", cut.Markup);
-    }
-
-    [Fact]
-    public void BuildCsv_Produces_Csv_String()
-    {
-        SetupServices();
-
-        var method = typeof(MetricsBeta).GetMethod("BuildCsv", BindingFlags.NonPublic | BindingFlags.Static)!;
-        var periodType = typeof(MetricsBeta).GetNestedType("PeriodMetrics", BindingFlags.NonPublic)!;
-        var array = Array.CreateInstance(periodType, 1);
-        var period = Activator.CreateInstance(periodType)!;
-        periodType.GetProperty("End")!.SetValue(period, new DateTime(2024, 1, 1));
-        periodType.GetProperty("AvgLeadTime")!.SetValue(period, 1.0);
-        periodType.GetProperty("AvgCycleTime")!.SetValue(period, 2.0);
-        periodType.GetProperty("Throughput")!.SetValue(period, 3);
-        periodType.GetProperty("Velocity")!.SetValue(period, 4.0);
-        array.SetValue(period, 0);
-
-        var csv = (string)method.Invoke(null, new object?[] { array })!;
-
-        Assert.Contains("Velocity", csv);
-        Assert.Contains("4.0", csv);
-    }
-
-    [Fact]
-    public void BuildPrompt_Includes_Metrics()
-    {
-        SetupServices();
-
-        var method = typeof(MetricsBeta).GetMethod("BuildPrompt", BindingFlags.NonPublic | BindingFlags.Static)!;
-        var periodType = typeof(MetricsBeta).GetNestedType("PeriodMetrics", BindingFlags.NonPublic)!;
-        var array = Array.CreateInstance(periodType, 1);
-        var period = Activator.CreateInstance(periodType)!;
-        periodType.GetProperty("End")!.SetValue(period, new DateTime(2024, 1, 1));
-        periodType.GetProperty("AvgLeadTime")!.SetValue(period, 1.2);
-        periodType.GetProperty("AvgCycleTime")!.SetValue(period, 2.3);
-        periodType.GetProperty("Throughput")!.SetValue(period, 3);
-        periodType.GetProperty("Velocity")!.SetValue(period, 4.5);
-        periodType.GetProperty("AvgWip")!.SetValue(period, 1.5);
-        periodType.GetProperty("SprintEfficiency")!.SetValue(period, 80.0);
-        array.SetValue(period, 0);
-
-        var prompt = (string)method.Invoke(null, new object?[] { array, OutputFormat.Markdown })!;
-
-        Assert.Contains("Agile Delivery Metrics Report Template", prompt);
-        Assert.Contains("\"end\":\"2024-01-01\"", prompt);
-        Assert.Contains("\"leadTime\":1.2", prompt);
-        Assert.Contains("\"velocity\":4.5", prompt);
-        Assert.Contains("\"avgWip\":1.5", prompt);
-        Assert.Contains("\"sprintEfficiency\":80", prompt);
-    }
-
-    [Fact]
-    public void ComputeBurnUp_DoesNot_Aggregate_Labels()
-    {
-        SetupServices();
-
-        var metrics = new TestMetricsBeta();
-        var type = typeof(MetricsBeta);
-        type.GetField("_additionalPoints", BindingFlags.NonPublic | BindingFlags.Instance)!.SetValue(metrics, (double?)10);
-        type.GetField("_efficiency", BindingFlags.NonPublic | BindingFlags.Instance)!.SetValue(metrics, (double?)80);
-        type.GetField("_errorRange", BindingFlags.NonPublic | BindingFlags.Instance)!.SetValue(metrics, (double?)10);
-        var compute = type.GetMethod("ComputeBurnUp", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        var labelsField = type.GetField("_burnLabels", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        var items = new List<StoryMetric>
-        {
-            new() { CreatedDate = DateTime.Today.AddDays(-3), ActivatedDate = DateTime.Today.AddDays(-2), ClosedDate = DateTime.Today.AddDays(-1), StoryPoints = 3 },
-            new() { CreatedDate = DateTime.Today.AddDays(-2), ActivatedDate = DateTime.Today.AddDays(-1), ClosedDate = DateTime.Today, StoryPoints = 2 }
-        };
-        compute.Invoke(metrics, new object?[] { items });
-
-        var labels = (string[])labelsField.GetValue(metrics)!;
-        Assert.All(labels, l => Assert.False(string.IsNullOrEmpty(l)));
-    }
-
-    [Fact]
-    public void ComputeFlow_DoesNot_Aggregate_Labels()
-    {
-        SetupServices();
-
-        var metrics = new TestMetricsBeta();
-        var type = typeof(MetricsBeta);
-        var compute = type.GetMethod("ComputeFlow", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        var labelsField = type.GetField("_flowLabels", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        var items = new List<StoryMetric>
-        {
-            new() { CreatedDate = DateTime.Today.AddDays(-3), ActivatedDate = DateTime.Today.AddDays(-2), ClosedDate = DateTime.Today.AddDays(-1) },
-            new() { CreatedDate = DateTime.Today.AddDays(-2), ActivatedDate = DateTime.Today.AddDays(-1), ClosedDate = DateTime.Today }
-        };
-        compute.Invoke(metrics, new object?[] { items });
-
-        var labels = (string[])labelsField.GetValue(metrics)!;
-        Assert.All(labels, l => Assert.False(string.IsNullOrEmpty(l)));
-    }
-
-
-    private class TestMetricsBeta : MetricsBeta
-    {
-        protected override Task OnInitializedAsync() => Task.CompletedTask;
-    }
-}
