+++ src/DevOpsAssistant/DevOpsAssistant.Tests/Pages/MetricsPageTests.cs
@@ -73,29 +73,7 @@ public class MetricsPageTests : ComponentTestBase
     }
 
     [Fact]
-    public void Fortnight_Mode_Uses_Two_Week_Period()
-    {
-        SetupServices();
-
-        var metrics = new TestMetrics();
-        var type = typeof(Metrics);
-        var compute = type.GetMethod("ComputePeriods", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        type.GetField("_mode", BindingFlags.NonPublic | BindingFlags.Instance)!.SetValue(metrics, AggregateMode.Fortnight);
-        type.GetField("_startDate", BindingFlags.NonPublic | BindingFlags.Instance)!.SetValue(metrics, DateTime.Today);
-        compute.Invoke(metrics, new object?[] { new List<StoryMetric>(), null });
-
-        var periodsField = type.GetField("_periods", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        var list = (System.Collections.IList)periodsField.GetValue(metrics)!;
-        Assert.True(list.Count > 0);
-        var first = list[0];
-        Assert.NotNull(first);
-        var start = (DateTime)first.GetType().GetProperty("Start")!.GetValue(first)!;
-        var end = (DateTime)first.GetType().GetProperty("End")!.GetValue(first)!;
-        Assert.Equal(13, (end - start).TotalDays);
-    }
-
-    [Fact]
-    public void ComputeBurnUp_Produces_Data()
+    public void ComputeBurnUp_DoesNot_Aggregate_Labels()
     {
         SetupServices();
 
@@ -105,7 +83,7 @@ public class MetricsPageTests : ComponentTestBase
         type.GetField("_efficiency", BindingFlags.NonPublic | BindingFlags.Instance)!.SetValue(metrics, (double?)80);
         type.GetField("_errorRange", BindingFlags.NonPublic | BindingFlags.Instance)!.SetValue(metrics, (double?)10);
         var compute = type.GetMethod("ComputeBurnUp", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        var seriesField = type.GetField("_burnSeries", BindingFlags.NonPublic | BindingFlags.Instance)!;
+        var labelsField = type.GetField("_burnLabels", BindingFlags.NonPublic | BindingFlags.Instance)!;
         var items = new List<StoryMetric>
         {
             new() { CreatedDate = DateTime.Today.AddDays(-3), ActivatedDate = DateTime.Today.AddDays(-2), ClosedDate = DateTime.Today.AddDays(-1), StoryPoints = 3 },
@@ -113,19 +91,18 @@ public class MetricsPageTests : ComponentTestBase
         };
         compute.Invoke(metrics, new object?[] { items });
 
-        var list = (IList<ChartSeries>)seriesField.GetValue(metrics)!;
-        Assert.True(list.Count > 0);
+        var labels = (string[])labelsField.GetValue(metrics)!;
+        Assert.All(labels, l => Assert.False(string.IsNullOrEmpty(l)));
     }
 
     [Fact]
-    public void ComputeFlow_Produces_Data()
+    public void ComputeFlow_DoesNot_Aggregate_Labels()
     {
         SetupServices();
 
         var metrics = new TestMetrics();
         var type = typeof(Metrics);
         var compute = type.GetMethod("ComputeFlow", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        var seriesField = type.GetField("_flowSeries", BindingFlags.NonPublic | BindingFlags.Instance)!;
         var labelsField = type.GetField("_flowLabels", BindingFlags.NonPublic | BindingFlags.Instance)!;
         var items = new List<StoryMetric>
         {
@@ -134,37 +111,10 @@ public class MetricsPageTests : ComponentTestBase
         };
         compute.Invoke(metrics, new object?[] { items });
 
-        var series = (IList<ChartSeries>)seriesField.GetValue(metrics)!;
         var labels = (string[])labelsField.GetValue(metrics)!;
-        Assert.True(series.Count > 0);
-        Assert.True(labels.Length > 0);
+        Assert.All(labels, l => Assert.False(string.IsNullOrEmpty(l)));
     }
 
-    [Fact]
-    public void ComputeOverallSprintEfficiency_Respects_Toggle()
-    {
-        SetupServices();
-
-        var metrics = new TestMetrics();
-        var type = typeof(Metrics);
-        type.GetField("_tag", BindingFlags.NonPublic | BindingFlags.Instance)!.SetValue(metrics, "Tech");
-        var compute = type.GetMethod("ComputeOverallSprintEfficiency", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        var modeField = type.GetField("_velocityMode", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        modeField.SetValue(metrics, VelocityMode.StoryPoints);
-        var toggleField = type.GetField("_tagMarksSprint", BindingFlags.NonPublic | BindingFlags.Instance)!;
-        var items = new List<StoryMetric>
-        {
-            new() { StoryPoints = 5, Tags = new[] { "Tech" } },
-            new() { StoryPoints = 3, Tags = new[] { "Other" } }
-        };
-
-        var result = (double)compute.Invoke(metrics, new object?[] { items })!;
-        Assert.Equal(37.5, result);
-
-        toggleField.SetValue(metrics, true);
-        result = (double)compute.Invoke(metrics, new object?[] { items })!;
-        Assert.Equal(62.5, result);
-    }
 
     private class TestMetrics : Metrics
     {
