+++ src/DevOpsAssistant/DevOpsAssistant/Components/ApexChartCard.razor
@@ -22,7 +22,8 @@
         <SimpleApexChart SeriesType="@SeriesType"
                          Series="Series"
                          Options="Options"
-                         FormatYAxisLabel="FormatYAxisLabel" />
+                         FormatYAxisLabel="FormatYAxisLabel"
+                         Class="responsive-chart" />
     </MudStack>
 </MudPaper>
