+++ src/DevOpsAssistant/DevOpsAssistant/Components/ApexChartDialog.razor
@@ -13,7 +13,8 @@
         <SimpleApexChart SeriesType="@SeriesType"
                          Series="Series"
                          Options="Options"
-                         FormatYAxisLabel="FormatYAxisLabel" />
+                         FormatYAxisLabel="FormatYAxisLabel"
+                         Class="responsive-chart" />
     </DialogContent>
     <DialogActions>
         <MudButton OnClick="Close" Color="MudBlazor.Color.Primary">@L["Close"]</MudButton>
