+++ src/DevOpsAssistant/DevOpsAssistant/Layout/MainLayout.razor
@@ -53,7 +53,6 @@
                 </MudNavGroup>
                 <MudNavGroup Title="@L["Reports"]" Expanded="true">
                         <MudNavLink Href="@($"projects/{_selectedProject}/metrics")" Icon="@Icons.Material.Filled.Insights" Disabled="@IsProjectInvalid">@L["Metrics"]</MudNavLink>
-                        <MudNavLink Href="@($"projects/{_selectedProject}/metrics-beta")" Icon="@Icons.Material.Filled.Insights" Disabled="@IsProjectInvalid">@L["MetricsBeta"]</MudNavLink>
                         <MudNavLink Href="@($"projects/{_selectedProject}/branch-health")" Icon="@Icons.Material.Filled.Source" Disabled="@IsProjectInvalid">@L["BranchHealth"]</MudNavLink>
                 </MudNavGroup>
                 <MudNavLink Href="@($"/projects/{_selectedProject}/settings")" Icon="@Icons.Material.Filled.Settings">@L["ProjectSettings"]</MudNavLink>
