+++ src/DevOpsAssistant/DevOpsAssistant/Layout/MainLayout.resx
@@ -27,9 +27,6 @@
   <data name="Metrics" xml:space="preserve">
     <value>Metrics</value>
   </data>
-  <data name="MetricsBeta" xml:space="preserve">
-    <value>Metrics (beta)</value>
-  </data>
   <data name="RequirementPlanner" xml:space="preserve">
     <value>Requirement Planner</value>
   </data>
