--- src/DevOpsAssistant/DevOpsAssistant/Pages/MetricsBeta.es.resx
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<root>
-  <resheader name="resmimetype">
-    <value>text/microsoft-resx</value>
-  </resheader>
-  <resheader name="version">
-    <value>2.0</value>
-  </resheader>
-  <resheader name="reader">
-    <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=8.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </resheader>
-  <resheader name="writer">
-    <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=8.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </resheader>
-  <data name="VelocityTooltip" xml:space="preserve">
-    <value>Elija qué campo usar para el cálculo de velocidad</value>
-  </data>
-  <data name="PageHeading" xml:space="preserve">
-    <value>Métricas (beta)</value>
-  </data>
-  <data name="OptionsHeading" xml:space="preserve">
-    <value>Opciones</value>
-  </data>
-  <data name="TableHeading" xml:space="preserve">
-    <value>Rendimiento y tiempos</value>
-  </data>
-  <data name="ChartsHeading" xml:space="preserve">
-    <value>Gráficos</value>
-  </data>
-  <data name="GeneratePrompt" xml:space="preserve">
-    <value>Generar mensaje de coaching</value>
-  </data>
-  <data name="CopyToast" xml:space="preserve">
-    <value>Copiado al portapapeles</value>
-  </data>
-  <data name="SprintEfficiencyOptionsHeading" xml:space="preserve">
-    <value>Opciones de eficiencia de sprint</value>
-  </data>
-  <data name="AvgWip" xml:space="preserve">
-    <value>WIP medio</value>
-  </data>
-  <data name="NonSprintTag" xml:space="preserve">
-    <value>Etiqueta</value>
-  </data>
-  <data name="TagMarksSprint" xml:space="preserve">
-    <value>La etiqueta indica trabajo de sprint</value>
-  </data>
-  <data name="UseSprintEff" xml:space="preserve">
-    <value>Usar eficiencia de sprint</value>
-  </data>
-  <data name="SprintEff" xml:space="preserve">
-    <value>% Sprint</value>
-  </data>
-  <data name="BurnUp" xml:space="preserve">
-    <value>Burn Up</value>
-  </data>
-  <data name="Flow" xml:space="preserve">
-    <value>Flujo acumulado</value>
-  </data>
-  <data name="ExpandChart" xml:space="preserve">
-    <value>Ampliar gráfico</value>
-  </data>
-  <data name="ProjectionOptions" xml:space="preserve">
-    <value>Opciones de proyección</value>
-  </data>
-  <data name="ProjectionHeading" xml:space="preserve">
-    <value>Proyección</value>
-  </data>
-  <data name="ProjectedDays" xml:space="preserve">
-    <value>Días restantes</value>
-  </data>
-  <data name="CompletionRange" xml:space="preserve">
-    <value>Rango de finalización</value>
-  </data>
-</root>
