--- src/DevOpsAssistant/DevOpsAssistant/Pages/MetricsBeta.resx
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<root>
-  <resheader name="resmimetype">
-    <value>text/microsoft-resx</value>
-  </resheader>
-  <resheader name="version">
-    <value>2.0</value>
-  </resheader>
-  <resheader name="reader">
-    <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=8.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </resheader>
-  <resheader name="writer">
-    <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=8.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </resheader>
-  <data name="VelocityTooltip" xml:space="preserve">
-    <value>Choose which field to use for velocity calculations</value>
-  </data>
-  <data name="PageHeading" xml:space="preserve">
-    <value>Metrics (beta)</value>
-  </data>
-  <data name="OptionsHeading" xml:space="preserve">
-    <value>Options</value>
-  </data>
-  <data name="EndDate" xml:space="preserve">
-    <value>End Date</value>
-  </data>
-  <data name="TableHeading" xml:space="preserve">
-    <value>Throughput and Lead Times</value>
-  </data>
-  <data name="ChartsHeading" xml:space="preserve">
-    <value>Charts</value>
-  </data>
-  <data name="GeneratePrompt" xml:space="preserve">
-    <value>Generate Coaching Prompt</value>
-  </data>
-  <data name="CopyToast" xml:space="preserve">
-    <value>Copied to clipboard</value>
-  </data>
-  <data name="SprintEfficiencyOptionsHeading" xml:space="preserve">
-    <value>Sprint Efficiency Options</value>
-  </data>
-  <data name="AvgWip" xml:space="preserve">
-    <value>Avg WIP</value>
-  </data>
-  <data name="NonSprintTag" xml:space="preserve">
-    <value>Tag</value>
-  </data>
-  <data name="TagMarksSprint" xml:space="preserve">
-    <value>Tag Marks Sprint Work</value>
-  </data>
-  <data name="UseSprintEff" xml:space="preserve">
-    <value>Use Sprint Efficiency</value>
-  </data>
-  <data name="SprintEff" xml:space="preserve">
-    <value>Sprint %</value>
-  </data>
-  <data name="BurnUp" xml:space="preserve">
-    <value>Burn Up</value>
-  </data>
-  <data name="Flow" xml:space="preserve">
-    <value>Cumulative Flow</value>
-  </data>
-  <data name="ExpandChart" xml:space="preserve">
-    <value>Expand Chart</value>
-  </data>
-  <data name="ProjectionOptions" xml:space="preserve">
-    <value>Projection Options</value>
-  </data>
-  <data name="ProjectionHeading" xml:space="preserve">
-    <value>Projection</value>
-  </data>
-  <data name="ProjectedDays" xml:space="preserve">
-    <value>Days Remaining</value>
-  </data>
-  <data name="CompletionRange" xml:space="preserve">
-    <value>Completion Range</value>
-  </data>
-</root>
