+++ src/DevOpsAssistant/DevOpsAssistant/wwwroot/css/app.css
@@ -188,9 +188,8 @@ body {
 /* Responsive chart container */
 .responsive-chart {
     width: 100%;
-    height: auto;
+    height: auto !important;
     aspect-ratio: 2 / 1;
-    max-height: 600px;
     padding-bottom: 3rem;
 }
